+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "homepage": "http://pixelzstudio.com",
   "scripts": {
-    "start" : "grunt"
+    "start" : "gulp"
   },
   "keywords": [
     "express",
