+++ src/client/routes/Editor/components/Range.js
@@ -78,7 +78,7 @@ obj.render = function() {
   if (this.props.large) {
     return <input className='range-large' type="range" max={this.props.max} min={this.props.min} value={this.props.value} onChange={this.onChange}/>;
   }
-  return <div className='range-small' onWheel={this.onWheel} style={{backgroundImage : this.getBackground()}} onMouseDown={this.onMouseDown} onS>
+  return <div className='range-small' onWheel={this.onWheel} style={{backgroundImage : this.getBackground()}} onMouseDown={this.onMouseDown}>
     <input value={this.props.value} onChange={this.onChange}/>
   </div>;
 };
